requirements.txt
streamlit==1.29.0
openai==1.3.7
python-dotenv==1.0.0
youtube-transcript-api==0.6.1
requests==2.31.0
beautifulsoup4==4.12.2
deepgram-sdk==3.7.5
python-slugify==8.0.1
httpx==0.25.2 